fixed test in duration
