--- CSL.Tests/UnitTest1.cs
@@ -1,58 +1,0 @@
-namespace CSL.Tests;
-
-using CSL;
-using NUnit.Framework;
-using Antlr4.Runtime;
-using System;
-using System.IO;
-
-[TestFixture]
-public class TestDuration
-{
-    [SetUp]
-    public void Setup()
-    {
-    }
-
-    [TestCase("10 min", 10 * Duration.MinuteFactor)]
-    [TestCase("5 h", 5 * Duration.HourFactor)]
-    [TestCase("2 d", 2 * Duration.DayFactor)]
-    [TestCase("3 w", 3 * Duration.WeekFactor)]
-    public void TestLiteralsMinuteParam(string input, int minutes)
-    {
-        var stream = CharStreams.fromString(input);
-
-        var lexer = new CSLLexer(stream);
-        var tokens = new CommonTokenStream(lexer);
-        var parser = new CSLParser(tokens);
-
-        var tree = parser.prog();
-
-        // Create and use the visitor
-        var visitor = new DurationVisitor();
-        var result = visitor.Visit(tree);
-
-        Assert.That(result, Is.Not.Null);
-        Assert.That(result.Minutes, Is.EqualTo(minutes));
-    }
-
-    [TestCase("2 mth", 2 * Duration.MonthFactor)]
-    [TestCase("1 y", 1 * Duration.YearFactor)]
-    public void TestLiteralMonthParam(string input, int months)
-    {
-        var stream = CharStreams.fromString(input);
-
-        var lexer = new CSLLexer(stream);
-        var tokens = new CommonTokenStream(lexer);
-        var parser = new CSLParser(tokens);
-
-        var tree = parser.prog();
-
-        // Create and use the visitor
-        var visitor = new DurationVisitor();
-        var result = visitor.Visit(tree);
-
-        Assert.That(result, Is.Not.Null);
-        Assert.That(result.Months, Is.EqualTo(months));
-    }
-}
